+++ fluxo-kmp-conf/src/main/kotlin/fluxo/conf/pub/SetupVanniktechPublication.kt
@@ -9,6 +9,7 @@ import com.vanniktech.maven.publish.JavaPlatform
 import com.vanniktech.maven.publish.KotlinJvm
 import com.vanniktech.maven.publish.KotlinMultiplatform
 import com.vanniktech.maven.publish.MavenPublishBaseExtension
+import com.vanniktech.maven.publish.SonatypeHost
 import com.vanniktech.maven.publish.VersionCatalog
 import envOrPropValue
 import fluxo.conf.data.BuildConstants.GRADLE_PLUGIN_PUBLISH_PLUGIN_ID
@@ -128,7 +129,10 @@ internal fun MavenPublishBaseExtension.setupVanniktechPublication(
 
         ConfigurationType.IDEA_PLUGIN -> {}
 
-        else -> publishToMavenCentral(config.sonatypeHost, automaticRelease = true)
+        else -> publishToMavenCentral(
+            host = config.sonatypeHost?.let { it as SonatypeHost } ?: SonatypeHost.DEFAULT,
+            automaticRelease = true,
+        )
     }
 
     // Log publications